Guess the number
A simple mini game where a random number is generated and the user is asked to guess the number. After guessing, the user is given feedback such as whether the number is greater or lesser than the number. The program terminates when the user guesses the correct number. The score of the player is based on the number of guesses. 
The score is calculated as (100 - number of guesses)